LayoutTests/imported/w3c/web-platform-tests/selection/selection-range-in-shadow-after-the-shadow-removed.tentative_mode=open-expected.txt


FAIL Selection range in shadow should not be as a selection range after the host is removed assert_equals: startContainer should not be changed expected Text node "ABC" but got Element node <div id="container"></div>
